More emojis and !deep command

+++ modules/checkforcode.js
@@ -1,6 +1,6 @@
 /*  checkforcode.js by David Jerome @GlassToeStudio - GlassToeStudio@gmail.com
 
-    14 July, 2017 
+    14 July, 2017
     https://github.com/GlassToeStudio
     http://glasstoestudio.weebly.com/
     https://twitter.com/GlassToeStudio
@@ -8,14 +8,14 @@
     ------------------------------------------------------------------------
     Systematically search through Discord comments to find unformatted Code.
 
-        * Search is based in chars in codeElements array. 
+        * Search is based in chars in codeElements array.
             * Default:  [';', '{', '}', ')', '[', ']', '>']
-        * Bot will parse the code line by line searching for 
+        * Bot will parse the code line by line searching for
         * code elements and keep track of which line are code
-        * and which are plain text. 
+        * and which are plain text.
         * The bot will do his best to only format a true code
         * block, leaving the plain text alone. One complete
-        * the bot will add code block formatting around the 
+        * the bot will add code block formatting around the
         * code block, with the current code Lang 'csharp'.
         * The message will be posted anew as formatted code
         * and, if possible, the old message will be deleted.
@@ -23,7 +23,7 @@
         * If the unformatted code is posted in a channel with
         * 'help' in the tile, then the new message is posted
         * there.
-        * If the message is posted in any other channel, the 
+        * If the message is posted in any other channel, the
         * new message will be posted in #programing_help (if it exists)
         * If there is no programing_help channel, the message is
         * posted in the original channel.
@@ -69,10 +69,10 @@
     https://support.discordapp.com/hc/en-us/articles/210298617-Markdown-Text-101-Chat-Formatting-Bold-Italic-Underline-?page=4
 */
 
-/*jshint esversion: 6 */
-/*jshint asi: true */
+/* jshint esversion: 6 */
+/* jshint asi: true */
 
-const Discord = require('discord.js');
+const Discord = require('discord.js')
 const GlassBot = require('../bot.js')
 const bot = GlassBot.bot
 const config = GlassBot.config
@@ -93,42 +93,39 @@ var selfDestructIn = 5
 
 // Lets begin
 bot.on('message', message => {
-    if (message.content.length > 1900) return
-    if (message.author.bot) {
+  if (message.content.length > 1900) return
+  if (message.author.bot) {
         // Self-destruct message
-        if (message.content.includes('Your unformatted code')) {
-            let usr = message.mentions.users.array()[0]
-            let chnl = (message.guild.channels.find("name", "programing_help") != null) ?
-                message.guild.channels.find("name", "programing_help") :
-                message.channel
-            callNTimes(5, 1000, EditBotMessage, message, chnl, usr)
-        }
-        return
+    if (message.content.includes('Your unformatted code')) {
+      let usr = message.mentions.users.array()[0]
+      let chnl = (message.guild.channels.find('name', 'programing_help') !== null)
+                ? message.guild.channels.find('name', 'programing_help')
+                : message.channel
+      callNTimes(5, 1000, EditBotMessage, message, chnl, usr)
     }
-    ParseMessage(message)
     return
-});
+  }
+  ParseMessage(message)
+})
 
 /**
- * Loop through each line in message and check for 
+ * Loop through each line in message and check for
  * code-like characters. If code formatting is found
  * return, else keep checking.
  * @param  {string[]} lines
  * @param  {string[]} message
  */
-function ParseMessage(message) {
-    InitVariables()
-    let lines = message.content.split('\n')
-    for (let i = 0; i < lines.length; i++) {
-        if (lines[i].search(formatBlock) >= 0) {
-            isFormatted = true
-            return
-        } else
-            FindCodeElements(i, lines[i], lines)
-    }
+function ParseMessage (message) {
+  InitVariables()
+  let lines = message.content.split('\n')
+  for (let i = 0; i < lines.length; i++) {
+    if (lines[i].search(formatBlock) >= 0) {
+      isFormatted = true
+      return
+    } else { FindCodeElements(i, lines[i], lines) }
+  }
 
-    CheckMessage(lines, message)
-    return
+  CheckMessage(lines, message)
 }
 
 /**
@@ -136,12 +133,11 @@ function ParseMessage(message) {
  * @param  {string[]} lines
  * @param  {string[]} message
  */
-function CheckMessage(lines, message) {
-    if (IsBadCode() && !isFormatted) {
-        lines[lastLine] = FormatLastLine(lines[lastLine])
-        CreateNewMessage(lines, message)
-    }
-    return
+function CheckMessage (lines, message) {
+  if (IsBadCode() && !isFormatted) {
+    lines[lastLine] = FormatLastLine(lines[lastLine])
+    CreateNewMessage(lines, message)
+  }
 }
 
 /**
@@ -150,21 +146,20 @@ function CheckMessage(lines, message) {
  * @param  {string} line
  * @param  {string[]} lines
  */
-function FindCodeElements(index, line, lines) {
-    let lineLength = line.length - 1
-    for (let i = 0; i < codeElements.length; i++) {
-        if (line.charAt(lineLength).valueOf() == codeElements[i].valueOf()) {
-            if (!hasFirstLine) {
-                lines[index] = FormatFirstLine(line)
-                return
-            } else {
-                lastLine = index
-                totalLinesOfCode += 1
-                return
-            }
-        }
+function FindCodeElements (index, line, lines) {
+  let lineLength = line.length - 1
+  for (let i = 0; i < codeElements.length; i++) {
+    if (line.charAt(lineLength).valueOf() === codeElements[i].valueOf()) {
+      if (!hasFirstLine) {
+        lines[index] = FormatFirstLine(line)
+        return
+      } else {
+        lastLine = index
+        totalLinesOfCode += 1
+        return
+      }
     }
-    return
+  }
 }
 
 /**
@@ -172,12 +167,11 @@ function FindCodeElements(index, line, lines) {
  * @param  {string[]} lines
  * @param  {string[]} message
  */
-function CreateNewMessage(lines, message) {
-    let newMessage = ""
-    for (let j = 0; j < lines.length; j++)
-        newMessage += lines[j] + '\n'
+function CreateNewMessage (lines, message) {
+  let newMessage = ''
+  for (let j = 0; j < lines.length; j++) { newMessage += lines[j] + '\n' }
 
-    PostNewMessage(message, newMessage)
+  PostNewMessage(message, newMessage)
 }
 
 /**
@@ -185,76 +179,73 @@ function CreateNewMessage(lines, message) {
  * @param  {string[]} message
  * @param  {string} newMessage
  */
-function PostNewMessage(message, newMessage) {
-    let channel = message.guild.channels.find("name", "programing_help")
-    let isHelp = message.channel.name.indexOf('help') > 0
+function PostNewMessage (message, newMessage) {
+  let channel = message.guild.channels.find('name', 'programing_help')
+  let isHelp = message.channel.name.indexOf('help') > 0
         // Move to new channel
-    if (channel != null && channel != message.channel && !isHelp) {
+  if (channel !== null && channel !== message.channel && !isHelp) {
         // TODO: Would like to add some color to this message
-        message.reply(':nerd: __`Your unformatted code has been formatted and moved to`__ ' + channel + '. :nerd:' +
+    message.reply(':nerd: __`Your unformatted code has been formatted and moved to`__ ' + channel + '. :nerd:' +
             '\n\t*This message will self-destruct in ' + selfDestructIn + ' seconds*')
-        channel.send(message.author + ', **★★ I have formatted your code and placed it here. Good Luck! ★★** ')
-        channel.send(newMessage);
+    channel.send(message.author + ', **★★ I have formatted your code and placed it here. Good Luck! ★★** ')
+    channel.send(newMessage)
         // post is same channel
-    } else {
-        message.channel.send(message.author + ' **★★ I see you forgot to format your code... Let me help you. ★★** ')
-        message.channel.send(newMessage)
-    }
+  } else {
+    message.channel.send(message.author + ' **★★ I see you forgot to format your code... Let me help you. ★★** ')
+    message.channel.send(newMessage)
+  }
 
-    DeleteOldMessage(message)
+  DeleteOldMessage(message)
 }
 
 /**
  * Deletes the old unformatted message if bot has permission
  * @param  {string[]} message
  */
-function DeleteOldMessage(message) {
-    let managePerms = message.guild.member(bot.user).hasPermission('MANAGE_MESSAGES')
-    if (managePerms)
-        message.delete()
-    else
-        message.channel.send('**`Tell the server\'s owner to grant me permission to delete your old message, thank\'s`** :wink:')
+function DeleteOldMessage (message) {
+  let managePerms = message.guild.member(bot.user).hasPermission('MANAGE_MESSAGES')
+  if (managePerms) { message.delete() } else { message.channel.send('**`Tell the server\'s owner to grant me permission to delete your old message, thank\'s`** :wink:') }
 }
 
 /**
  *  Adds code formatting block start to the first line of code
  * @param  {string} firstLine
  */
-function FormatFirstLine(firstLine) {
+function FormatFirstLine (firstLine) {
     /*
     What if the first line of code has some regular text at the beginning?
         "Here is my code: public int myInt = 0;"
         "Here is my code" will also be formatted.
         We do not want this.
      */
-    hasFirstLine = true
-    return formatBlock + codeLang + '\n' + firstLine
+  hasFirstLine = true
+  return formatBlock + codeLang + '\n' + firstLine
 }
 
 /**
  * Add formatting code bock end to the last line of code
  * @param  {string} lastLine
  */
-function FormatLastLine(lastLine) {
-    return lastLine + '\n' + formatBlock
+function FormatLastLine (lastLine) {
+  return lastLine + '\n' + formatBlock
 }
 
 /**
  * If total line of code is greater than repostThreshold return true
  */
-function IsBadCode() {
-    return (totalLinesOfCode >= repostThreshold)
+function IsBadCode () {
+  return (totalLinesOfCode >= repostThreshold)
 }
 
 /**
  * Initialize variables
  */
-function InitVariables() {
-    isFormatted = false
-    hasFirstLine = false
-    lastLine = 0
-    totalLinesOfCode = 0
-    selfDestructIn = 5
+function InitVariables () {
+  isFormatted = false
+  hasFirstLine = false
+  lastLine = 0
+  totalLinesOfCode = 0
+  selfDestructIn = 5
 }
 
 // Bot self destruct message functions
@@ -266,8 +257,8 @@ function InitVariables() {
  * @param  {string} channel
  * @param  {number} t
  */
-function EditBotMessage(usr, message, channel, t) {
-    message.edit(usr + ', :nerd: __`Your unformatted code has been formatted and moved to`__ ' + channel + '. :nerd:' +
+function EditBotMessage (usr, message, channel, t) {
+  message.edit(usr + ', :nerd: __`Your unformatted code has been formatted and moved to`__ ' + channel + '. :nerd:' +
         '\n\t*This message will self-destruct in ' + t + ' seconds*')
 }
 
@@ -280,17 +271,17 @@ function EditBotMessage(usr, message, channel, t) {
  * @param  {string} chnl
  * @param  {string} usr
  */
-function callNTimes(n, time, fn, msg, chnl, usr) {
-    function callFn() {
-        if (--n < 1) {
-            usr = null
-            msg.delete()
+function callNTimes (n, time, fn, msg, chnl, usr) {
+  function callFn () {
+    if (--n < 1) {
+      usr = null
+      msg.delete()
                 .then(m => console.log(`Deleted message from ${m.author}`))
-                .catch(console.error);
-            return;
-        }
-        fn(usr, msg, chnl, n);
-        setTimeout(callFn, time);
+                .catch(console.error)
+      return
     }
-    setTimeout(callFn, time);
+    fn(usr, msg, chnl, n)
+    setTimeout(callFn, time)
+  }
+  setTimeout(callFn, time)
 }
